+++ src/editor/tests/editor.js
@@ -543,7 +543,7 @@ YUI.add('editor-tests', function(Y) {
 
             },
             error: { //These tests should error
-                'test: EditorSelection': (Y.UA.chrome),
+                'test: EditorSelection': (Y.UA.chrome || Y.UA.webkit),
                 test_selection_methods: (Y.UA.ie || Y.UA.webkit ? true : false),
                 test_execCommands: ((Y.UA.webkit || (Y.UA.ie && Y.UA.ie === 9) || Y.UA.chrome) ? true : false),
                 test_double_plug: true,